oem unlock: disable the ability to deny unlocking
even though there are clear and BIG warnings that users should toggle the "Allow the bootloader to be unlocked" only when the risks are clearly understood, people still bricked their devices in the past as they simply do not read - or understand the impact - or both.

to avoid this we force-disable the ability to do so. due to magisk people are still able to overwrite this but that requires at least manual interaction via ADB / terminal and so adds at least  another (little) barrier to avoid bricking the device.

Signed-off-by: steadfasterX <steadfasterX@users.noreply.github.com>